smol-agent/pyproject.toml
[project]
name = "smol-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.9.3",
    "duckduckgo-search>=8.1.1",
    "litellm>=1.80.9",
    "markdownify>=1.2.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "smolagents[toolkit,transformers]>=1.23.0",
]
